Create expense_tracker file, move code, and tests
